+++ ngClient/js/app/controllers/account/changeEmailController.js
@@ -17,7 +17,7 @@
         function _init() {
 
             // Generate test data if localhost
-            if ($location.$$host === 'localhost') {
+            if ($location.host() === 'localhost') {
                 var now = new Date();
                 var year = now.getFullYear();
                 var month = now.getMonth() + 1;